Add function remap_field_dryrun
The purpose of this function is to produce a target field with the correct
field name, dimension names, and dimension sizes, without actually having
to read the input MPAS field and perform the remapping.

It may be simpler to just implement this functionality in the normal remap_field_dryrun
routine with an optional argument.