Add field_list module to manage fields to be included or excluded from output
This module looks for two files: 'include_fields', which contains a list of fields
to be included in the interpolated output, and 'exclude_fields', which contains
a list of fields to be excluded from the output.

If only an 'include_fields' file is present, only fields in that list will be interpolated.
If both an 'include_fields' file and an 'exclude_fields' file are present, only
the 'include_fields' file will be considered.
If only and 'exclude_fields' file is present, all fields that do not appear in that list
will be interpolated.