Enable appending to an existing output file in file_output module
This commit adds support for two modes when opening an output file:
1) Any existing file can be clobbered
2) The existing file can be appended with new records

When opening an output file, the clobber mode is controlled by a new
optional argument to file_output_open, mode. Also, if present,
another new optional argument, nRecords, will return the number of
records already in the output file.

It is not possible to define new fields in an existing output file,
and trying to register a new field will result in an error.