Generalize target_mesh_type for 2-d arrays of latitude and longitude
Other applications may require 2-d arrays of latitude and longitude,
while 1-d arrays are still more efficient for regular lat-lon grids.
To generalize the 'lats' and 'lons' arrays while still maintaining
efficiency of storage in the current convert_mpas code, we add an
extra dimension to the 'lats' and 'lons' members of target_mesh_type,
and introduce a function that maps (ix,iy) indices for these arrays
into an effectively 1-d array as needed. For "effectively 1-d" arrays,
the 'irank' member of the target_mesh_type must be set to 0.