Allow optional 2-d lat,lon fields to be provided when setting up target mesh
New optional arguments, lat2d and lon2d, to the target_mesh_setup( ) routine
allow externally supplied, arbitrary grids of locations to be used for
the target mesh. When such arguments are supplied, the rank of the target
grid is set to 1 and the internal 'lats' and 'lons' arrays of the target mesh,
which point to 'lat2d' and 'lon2d', are only deallocated when the rank is 0,
ensuring that the target mesh does not free memory that may still be needed by
the driving code.

Note that the 'lat2d' and 'lon2d' fields should have units of radians.