Support compilation for old netCDF libraries without nf90_inq_varids
Newer versions of the netCDF4 library (perhaps newer than 4.2.0?) provide
a function, nf90_inq_varids, to return a list of all variable IDs in a file;
if we are using an older version of the netCDF library, we can apparently
assume that the variable IDs are numbered 1 through nVars.
See http://www.unidata.ucar.edu/software/netcdf/docs/tutorial_ncids.html
for more details.

When building convert_mpas with an older version of the netCDF libraries,
the pre-processing macro HAVE_NF90_INQ_VARIDS should be undefined when
compiling scan_input.F. By default, this macro is defined in FFLAGS in
the top-level Makefile.