Improve logic for handling input files with no unlimited dimension
When inquiring about the size of a non-existent unlimited dimension
through the nf90_inquire_dimension routine, some versions of the netCDF
library return a zero length, while others do not. This leads to
failures like the following when trying to convert, e.g., a mesh file:

   Remapping MPAS fields from file 'grid.nc'
   Error: Problems opening input file grid.nc
          This could result from an input file with no unlimited dimension.
  STOP 6

In order to better handle the behavior across different netCDF library
versions, we now check whether the unlimited dimension exists (based on
its ID), and if not, as long as the number of variables in the files is
non-zero, we assume at least one record.