Preliminary changes to transpose dimensions in output arrays
The changes in this commit cause the n-1 fastest varying dimensions
in the input array (e.g., nVertLevels in a field dimensioned (nVertLeve,nCells))
to become the slowest-varying dimensions in the output array, so that
nLon is fastest varying and nLat is second fastest varying.

The performance of this code is somewhat slower, since memory in
the output array is not accessed in an optimal sequence. But, the resulting
array order will allow one to plot vertical cross-sections with ncview.