Obtain nCells, nEdges, and nVertices for general meshes, not just global meshes
The logic for obtaining nCells, nEdges, and nVertices from the source mesh
previously read nCells as the dimension of the nEdgesOnCell field, then derived
nVertices as 2*(nCells-2) and nEdges as 3*(nCells-2). This works for global
meshes, but will give incorrect results for regional or doubly-periodic meshes.

Now, nCells, nEdges, and nVertices are all read directly from the input mesh
file to ensure they are correct for any input topology.