Implement support for remapping of limited-area MPAS domains
This commit introduces changes to allow the convert_mpas tool to process
limited-area MPAS meshes. Setting up the target grid with the 'target_domain'
file is done as for global meshes, and target grid points that are not covered
by the limited-area MPAS domain will receive the default netCDF fill values.
This fill value will also be indicated by the _FillValue attribute on output
variables

At present, values from the outermost layer of cells, edges, and vertices
will not be represented in the interpolated field. Target grid points must be
surrounded by valid cells, edges, and vertices in order to receive
 an interpolated value

Implementation notes:

* The outines nearest_cell and nearest_vertex will return an index of 0 if
  the path-based search for the nearest cell or vertex encounters a cell
  or vertex with neighbors having index 0 (which is used to indicate
  non-existent elements in a limited-area mesh)

* In future, nearest-neighbor interpolation could be modified to support
  mapping of outermost cells/edges/vertices with checks to verify that
  the nearest cell returned by a search actually contains the target location

* Internally, target grid points to which we will not interpolate may be
  identified as points for which all interpolation weights are zero