Optimize remapping weight calculation for limited-area meshes
When remapping fields from limited-area meshes, if there are target
grid points that fall outside the coverage of the limited-area mesh,
the searches for nearest cells, nearest edges, or nearest vertices
in the remap_info_setup(...) routine would restart the search from
element 1 after each failed search (i.e., a search that resulted in
a boundary element).

To reduce the time for searches after a failed search, we can start
from the last element in the mesh to contain a target grid point. This
is accomplished by setting the starting element for each new search
('last_idx' in the code) to the previous element only if the previous
search was successful.

The time to compute remapping weights when the target grid is completely
covered by the source mesh is not affected, and the time to compute
remapping weights when part of the target grid is not covered is now
much faster (comparable to the case where the target grid is covered).