Choose an "interior" element from which to begin nearest element search
When choosing an initial element (cell, edge, or vertex) index from which
we begin the very first search for the nearest element, always choose
an "interior" element (i.e., an element with no missing cell neighbors).

Previously, index 1 was chosen as the beginning of the first search. However,
if index 1 was on the boundary of a regional mesh, all subsequent searches
would fail, since, e.g., nearest_cell returns a zero index if any neighbor
of the search point is missing. The abbreviated loop, below, illustrates this
problem:

        last_idx = 1   ! This choice can potentially be problematic

        do iy=1,dst_mesh % nlat
        do ix=1,dst_mesh % nlon

            ! If last_idx is a boundary element, nearest_cell will fail
            idx = nearest_cell(..., last_idx, ...)

            ! If nearest_cell fails, we never change our search starting
            ! point, last_idx
            if (idx <= 0) then
                remap_info % nearestCell(ix, iy) = 0
            else
                remap_info % nearestCell(ix, iy) = idx
                last_idx = idx
            end if
        end do
        end do

As a solution, a new routine, interior_element, has been added to provide
a suitable, non-boundary starting element for the first search.