Generalize remapping code to work with CAM-MPAS or stand-alone MPAS output
While stand-alone MPAS output always contains the mesh element dimension
(nCells, nEdges, or nVertices) as the slowest-varying dimension, CAM-MPAS
output may contain the mesh element dimension as the fastest-varying dimension.
Furthermore, CAM may use both of the dimension names 'ncol' and 'nCells' to
refer to the cell dimension of fields.

To accommodate these different orderings and names of dimensions, this commit
introduces constants to indicate the mesh element type on which a field is
defined and the dimension ordering of that field, along with a new function to
return an appropriate constant given the list of dimension names for an input
(source) field. Masks are also provided to determine whether a field --
regardless of its dimension ordering -- is a cell-, vertex-, or edge-based
field, and whether a field uses CAM-MPAS or stand-alone MPAS dimension order.