Add pre-build check in top-level Makefile for the NetCDF Fortran 90 interface
The top-level Makefile now runs a check to detect whether the Fortran 90
interface to the NetCDF library is available; this is accomplished by compiling
a small test program that tries to use the nf90_open routine from the netcdf
module. If the compilation of the test program fails, an explanatory error
message is printed.