Add pre-build check in top-level Makefile for availability of nf90_inq_varids
The top-level Makefile now runs a check to determine whether the NetCDF
library provides the nf90_inq_varids routine. If the nf90_inq_varids routine
is available, -DHAVE_NF90_INQ_VARIDS is added to the list of Fortran
preprocessing flags. Note that the code already made use of the
HAVE_NF90_INQ_VARIDS macro, and the purpose of the changes in this commit
is to determine whether this macro should be defined.