Address netCDF shared library issues through the use of rpath
On NCAR's Derecho system, compiler wrappers generally handle the setting of
rpath for shared libraries in compiled executables. However, the Fortran
compiler used in the top-level Makefile is set to the actual compiler used to
build the netCDF library using `nf-config --fc`, thereby precluding the use of
the compiler wrapper. This leads to runtime issues in locating the netCDF shared
libraries, e.g.,

  ./convert_mpas: error while loading shared libraries: libnetcdff.so.7: cannot open shared object file: No such file or directory

As a workaround, the top-level Makefile now tries to extract linker rpath flags
from the output of `nf-config --flibs` and `nc-config --libs`. These flags are
stored in the RPATH_FLAGS make variable, which is then appended to the FCLIBS
variable.